requirements.txt
anyio==3.6.2
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cycler==0.11.0
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.86.0
fonttools==4.38.0
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.23.4
opencv-python==4.5.5.64
orjson==3.8.1
packaging==21.3
pandas==1.4.1
Pillow==9.0.0
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
redis==4.3.4
requests==2.28.1
scipy==1.9.3
seaborn==0.11.2
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
torch==1.10.2
torchvision==0.11.3
tqdm==4.63.0
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.18.1
websockets==10.4
wrapt==1.14.1
